refactor: timers proto packages
