+++ infrastructure/services/src/main/kotlin/io/timemates/backend/services/timers/TimersMapper.kt
@@ -2,7 +2,7 @@ package io.timemates.backend.services.timers
 
 import io.timemates.api.timers.requests.EditTimerInfoRequestOuterClass
 import io.timemates.api.timers.requests.EditTimerSettingsRequestOuterClass
-import io.timemates.api.timers.requests.members.invites.types.InviteOuterClass
+import io.timemates.api.timers.members.invites.types.InviteOuterClass
 import io.timemates.api.timers.types.TimerKt.state
 import io.timemates.api.timers.types.TimerOuterClass
 import io.timemates.api.timers.types.TimerOuterClass.Timer.State
