+++ infrastructure/services/src/main/kotlin/io/timemates/backend/services/timers/TimersService.kt
@@ -3,15 +3,13 @@ package io.timemates.backend.services.timers
 import com.google.protobuf.Empty
 import io.grpc.Status
 import io.grpc.StatusException
-import io.timemates.api.common.types.StatusOuterClass
 import io.timemates.api.timers.TimersServiceGrpcKt
+import io.timemates.api.timers.members.invites.requests.*
+import io.timemates.api.timers.members.requests.GetMembersRequestKt
+import io.timemates.api.timers.members.requests.GetMembersRequestOuterClass.GetMembersRequest
+import io.timemates.api.timers.members.requests.KickMemberRequestOuterClass
 import io.timemates.api.timers.requests.*
 import io.timemates.api.timers.requests.CreateTimerRequestOuterClass.CreateTimerRequest
-import io.timemates.api.timers.requests.members.GetMembersRequestKt
-import io.timemates.api.timers.requests.members.GetMembersRequestOuterClass
-import io.timemates.api.timers.requests.members.GetMembersRequestOuterClass.GetMembersRequest
-import io.timemates.api.timers.requests.members.KickMemberRequestOuterClass
-import io.timemates.api.timers.requests.members.invites.*
 import io.timemates.api.timers.types.TimerOuterClass
 import io.timemates.backend.common.types.value.Count
 import io.timemates.backend.pagination.PageToken
