+++ infrastructure/services/src/main/proto/io/timemates/api/timers/TimersService.proto
@@ -6,11 +6,11 @@ import "io/timemates/api/timers/requests/EditTimerSettingsRequest.proto";
 import "io/timemates/api/timers/requests/RemoveTimerRequest.proto";
 import "io/timemates/api/timers/requests/GetTimerRequest.proto";
 import "io/timemates/api/timers/requests/GetTimersRequest.proto";
-import "io/timemates/api/timers/requests/members/KickMemberRequest.proto";
-import "io/timemates/api/timers/requests/members/GetMembersRequest.proto";
-import "io/timemates/api/timers/requests/members/invites/CreateInviteRequest.proto";
-import "io/timemates/api/timers/requests/members/invites/GetInvitesRequest.proto";
-import "io/timemates/api/timers/requests/members/invites/RemoveInviteRequest.proto";
+import "io/timemates/api/timers/members/requests/KickMemberRequest.proto";
+import "io/timemates/api/timers/members/requests/GetMembersRequest.proto";
+import "io/timemates/api/timers/members/invites/requests/CreateInviteRequest.proto";
+import "io/timemates/api/timers/members/invites/requests/GetInvitesRequest.proto";
+import "io/timemates/api/timers/members/invites/requests/RemoveInviteRequest.proto";
 import "io/timemates/api/common/types/Status.proto";
 import "io/timemates/api/timers/types/Timer.proto";
 import "google/protobuf/empty.proto";
