+++ infrastructure/services/src/main/proto/io/timemates/api/timers/members/invites/requests/CreateInviteRequest.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 import "io/timemates/api/users/types/User.proto";
 
-option java_package = "io.timemates.api.timers.requests.members.invites";
+option java_package = "io.timemates.api.timers.members.invites.requests";
 
 message InviteMemberRequest {
   int64 timerId = 1;
