+++ infrastructure/services/src/main/proto/io/timemates/api/timers/members/invites/requests/GetInvitesRequest.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-import "io/timemates/api/timers/requests/members/invites/types/Invite.proto";
+import "io/timemates/api/timers/members/invites/types/Invite.proto";
 
-option java_package = "io.timemates.api.timers.requests.members.invites";
+option java_package = "io.timemates.api.timers.members.invites.requests";
 
 message GetInvitesRequest {
   int64 timerId = 1;
