+++ infrastructure/services/src/main/proto/io/timemates/api/timers/members/invites/requests/RemoveInviteRequest.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option java_package = "io.timemates.api.timers.requests.members.invites";
+option java_package = "io.timemates.api.timers.members.invites.requests";
 
 message RemoveInviteRequest {
   int64 timerId = 1;
