+++ infrastructure/services/src/main/proto/io/timemates/api/timers/members/requests/GetMembersRequest.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 import "io/timemates/api/users/types/User.proto";
 
-option java_package = "io.timemates.api.timers.requests.members";
+option java_package = "io.timemates.api.timers.members.requests";
 
 message GetMembersRequest {
   int64 timerId = 1;
